requirements.txt
python-dotenv
pytest
colorama==0.4.6
typer
openai